requirements.txt

anthropic==0.75.0
beautifulsoup4==4.12.2
ddgs==9.9.3
duckduckgo_search==8.1.1
langchain==1.2.0
langchain-anthropic==1.3.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.2.1
langchain-ollama==1.0.1
langchain-openai==1.1.3
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.0
lxml==6.0.2
openai==1.109.1
python-dotenv==1.0.0
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
streamlit==1.29.0


